Update CMake script. Fixed nearly all warnings. Support Qt4 and Qt5 f…

+++ coin-competitions/SIM-20020503/strutils.h
@@ -29,6 +29,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 \**************************************************************************/
+#include <cstdlib>
 
-char * strconcat(const char * array[], unsigned int &total);
+char * strconcat(const char * array[], size_t &total);
